+++ src/ScottPlot5/ScottPlot5 Tests/RenderTests/LegendTests.cs
@@ -14,10 +14,28 @@ internal class LegendTests
 
         plt.SaveTestImage(300, 200, "legend-default");
 
-        plt.Legend(true);
+        plt.Legend.IsVisible = true;
         plt.SaveTestImage(300, 200, "legend-enabled");
 
-        plt.Legend(false);
+        plt.Legend.IsVisible = false;
         plt.SaveTestImage(300, 200, "legend-disabled");
+    }
+
+    [Test]
+    public void Test_Legend_FontStyle()
+    {
+        Plot plt = new();
+
+        var sig1 = plt.Add.Signal(Generate.Sin());
+        var sig2 = plt.Add.Signal(Generate.Cos());
+
+        sig1.Label = "Sine";
+        sig2.Label = "Cosine";
+
+        plt.Legend.IsVisible = true;
+        plt.Legend.Font.Size = 26;
+        plt.Legend.Font.Color = Colors.Magenta;
+
+        plt.SaveTestImage();
     }
 }