+++ src/ScottPlot5/ScottPlot5/DataSources/ScatterSourceCoordinates.cs
@@ -14,6 +14,7 @@ public class ScatterSourceCoordinates : IScatterSource
 
     public IReadOnlyList<Coordinates> GetScatterPoints()
     {
+        // TODO: try to avoid calling this
         return Coordinates;
     }
 
@@ -32,5 +33,34 @@ public class ScatterSourceCoordinates : IScatterSource
     public CoordinateRange GetLimitsY()
     {
         return GetLimits().Rect.YRange;
+    }
+
+    public DataPoint GetNearest(Coordinates mouseLocation, RenderDetails renderInfo, float maxDistance = 15)
+    {
+        double maxDistanceSquared = maxDistance * maxDistance;
+        double closestDistanceSquared = double.PositiveInfinity;
+
+        int closestIndex = 0;
+        double closestX = double.PositiveInfinity;
+        double closestY = double.PositiveInfinity;
+
+        for (int i = 0; i < Coordinates.Count; i++)
+        {
+            double dX = (Coordinates[i].X - mouseLocation.X) * renderInfo.PxPerUnitX;
+            double dY = (Coordinates[i].Y - mouseLocation.Y) * renderInfo.PxPerUnitY;
+            double distanceSquared = dX * dX + dY * dY;
+
+            if (distanceSquared <= closestDistanceSquared)
+            {
+                closestDistanceSquared = distanceSquared;
+                closestX = Coordinates[i].X;
+                closestY = Coordinates[i].Y;
+                closestIndex = i;
+            }
+        }
+
+        return closestDistanceSquared <= maxDistanceSquared
+            ? new DataPoint(closestX, closestY, closestIndex)
+            : DataPoint.None;
     }
 }