+++ src/ScottPlot5/ScottPlot5/Interfaces/ILegend.cs
@@ -1,7 +1,0 @@
-﻿namespace ScottPlot;
-
-public interface ILegend
-{
-    bool IsVisible { get; set; }
-    void Render(RenderPack rp);
-}